EES-6914 tidy comments and components into separate files
