+++ src/explore-education-statistics-admin/src/pages/release/data/components/DataFileReplacementDifferences.tsx
@@ -1,34 +1,15 @@
-import { IndicatorSource } from '@admin/services/apiDataSetVersionService';
 import dataReplacementService, {
-  IndicatorMappingWithKey,
-  IndicatorsMappingsPlan,
-  MappingWithKey,
   PlanMappings,
-  UpdateMappingPayload,
   UpdateMappingPayloadMultiple,
 } from '@admin/services/dataReplacementService';
-import Button from '@common/components/Button';
-import ButtonGroup from '@common/components/ButtonGroup';
-import ButtonText from '@common/components/ButtonText';
-import { Form } from '@common/components/form';
-import FormFieldRadioSearchGroup from '@common/components/form/FormFieldRadioSearchGroup';
-import FormProvider from '@common/components/form/FormProvider';
-import { RadioOption } from '@common/components/form/FormRadioGroup';
-import LoadingSpinner from '@common/components/LoadingSpinner';
-import Modal from '@common/components/Modal';
-import SummaryList from '@common/components/SummaryList';
-import SummaryListItem from '@common/components/SummaryListItem';
 import Tag from '@common/components/Tag';
 import TagGroup from '@common/components/TagGroup';
 import VisuallyHidden from '@common/components/VisuallyHidden';
-import { Dictionary } from '@common/types';
-import Yup from '@common/validation/yup';
-import classNames from 'classnames';
 import mapValues from 'lodash/mapValues';
-import pickBy from 'lodash/pickBy';
 import startCase from 'lodash/startCase';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo } from 'react';
 import { useImmer } from 'use-immer';
+import DifferencesMappingTableRows from './DataFileReplacementDifferencesTableRows';
 
 interface Props {
   releaseVersionId: string;
@@ -37,8 +18,6 @@ interface Props {
   mapping: PlanMappings;
   reloadPlan: () => void;
 }
-
-type SourceItem = IndicatorSource /* | FilterSource | LocationSource */;
 
 export default function DataFileReplacementDifferences({
   releaseVersionId,
@@ -139,7 +118,6 @@ export default function DataFileReplacementDifferences({
         </caption>
         <thead>
           <VisuallyHidden as="tr">
-            {/* <th className="govuk-!-width-one-quarter">Item Type</th> */}
             <th className="govuk-!-width-one-quarter">Original Item</th>
             <th className="govuk-!-width-one-quarter">Mapping</th>
             <th className="govuk-!-text-align-right">Actions</th>
@@ -154,270 +132,5 @@ export default function DataFileReplacementDifferences({
         </tbody>
       </table>
     </>
-  );
-}
-
-/* =-=-=-=-=-=-=-= DifferencesMappingTableRows =-=-=-=-=-=-=-= */
-
-interface RowsProps {
-  itemType: 'indicator' | 'filter' | 'location';
-  mappings: IndicatorsMappingsPlan /* | FiltersMappingsPlan | LocationsMappingsPlan */;
-  onUpdate: (payload: UpdateMappingPayloadMultiple) => Promise<void>;
-}
-
-function DifferencesMappingTableRows({
-  itemType,
-  mappings: { candidates, mappings },
-  onUpdate,
-}: RowsProps) {
-  const [currentMappingItem, setCurrentMappingItem] = useState<
-    IndicatorMappingWithKey | undefined
-  >();
-  const [loadingMappings, setLoadingMappings] = useState<
-    Set<UpdateMappingPayload['sourceKey']>
-  >(new Set());
-
-  const handleModalMappingSubmit = useCallback(
-    (payload: UpdateMappingPayloadMultiple) => {
-      const updateMappingKeys = payload.map(({ sourceKey }) => sourceKey);
-      setLoadingMappings(prev => {
-        const next = new Set(prev);
-        updateMappingKeys.forEach(key => next.add(key));
-        return next;
-      });
-
-      onUpdate(payload).finally(() =>
-        setLoadingMappings(prev => {
-          const next = new Set(prev);
-          updateMappingKeys.forEach(key => next.delete(key));
-          return next;
-        }),
-      );
-    },
-    [],
-  );
-
-  const {
-    allCandidates,
-    unmappedCandidates,
-  }: {
-    allCandidates: Dictionary<SourceItem>;
-    unmappedCandidates: Dictionary<SourceItem>;
-  } = useMemo(() => {
-    // add key inside to each item
-    const allCandidatesWithKeys = mapValues(
-      candidates,
-      ({ label }, candidateKey) => ({ candidateKey, label }),
-    );
-
-    const unmappedCandidatesWithKeys = pickBy(
-      allCandidatesWithKeys,
-      ({ candidateKey }) =>
-        // filter out candidates that exist in mappings
-        !Object.values(mappings).some(mapping => {
-          return mapping.candidateKey === candidateKey;
-        }),
-    );
-
-    return {
-      allCandidates: allCandidatesWithKeys,
-      unmappedCandidates: unmappedCandidatesWithKeys,
-    };
-  }, [candidates, mappings]);
-
-  const mappingsToList: IndicatorMappingWithKey[] = Object.entries(mappings)
-    .map(([sourceKey, mapping]) => ({ ...mapping, sourceKey }))
-    .filter(({ type }) => type !== 'AutoSet');
-
-  return (
-    <>
-      {currentMappingItem && (
-        <DifferencesItemMappingModal
-          itemType={itemType}
-          allCandidateOptions={allCandidates}
-          unmappedCandidateOptions={unmappedCandidates}
-          currentItem={currentMappingItem}
-          onSubmit={handleModalMappingSubmit}
-          onClose={() => setCurrentMappingItem(undefined)}
-        />
-      )}
-
-      {mappingsToList.map((mapping, index) => {
-        const { source, sourceKey, type } = mapping;
-        const isUnset = type === 'Unset';
-        const itemCurrentMapping =
-          (mapping.candidateKey &&
-            allCandidates[mapping.candidateKey]?.label) ??
-          'No mapping';
-        const isLoading = loadingMappings.has(mapping.sourceKey);
-
-        return (
-          <tr
-            key={`mapping-${sourceKey}`}
-            className={classNames({
-              'rowHighlight--alert': isUnset,
-            })}
-          >
-            {/* <td>
-              <strong>
-                {index === 0 ? (
-                  startCase(itemType)
-                ) : (
-                  <VisuallyHidden>{itemType}</VisuallyHidden>
-                )}
-              </strong>
-            </td> */}
-            <td>{source.label}</td>
-            <td>
-              {isUnset ? (
-                <Tag colour="red">not present</Tag>
-              ) : (
-                itemCurrentMapping
-              )}
-            </td>
-            <td className="govuk-!-text-align-right">
-              <LoadingSpinner
-                loading={isLoading}
-                alert
-                hideText
-                inline
-                size="sm"
-                text={`Updating mapping for ${mapping.source.label}`}
-              >
-                {mapping.type === 'Unset' && (
-                  <ButtonText
-                    onClick={() =>
-                      handleModalMappingSubmit([
-                        {
-                          sourceKey: mapping.sourceKey,
-                          candidateKey: undefined, // no mapping
-                        },
-                      ])
-                    }
-                  >
-                    No mapping{' '}
-                    <VisuallyHidden>for {mapping.source.label}</VisuallyHidden>
-                  </ButtonText>
-                )}
-
-                <ButtonText
-                  className="govuk-!-margin-left-2"
-                  onClick={() => setCurrentMappingItem(mapping)}
-                >
-                  Map item{' '}
-                  <VisuallyHidden>{mapping.source.label}</VisuallyHidden>
-                </ButtonText>
-              </LoadingSpinner>
-            </td>
-          </tr>
-        );
-      })}
-    </>
-  );
-}
-
-interface ModalProps {
-  itemType: string;
-  allCandidateOptions: Dictionary<SourceItem>;
-  unmappedCandidateOptions: Dictionary<SourceItem>;
-  currentItem: MappingWithKey<SourceItem>;
-  onSubmit: (payload: UpdateMappingPayloadMultiple) => void;
-  onClose: () => void;
-}
-
-interface FormValues {
-  selectedCandidate: string;
-}
-
-function DifferencesItemMappingModal({
-  currentItem,
-  itemType,
-  allCandidateOptions,
-  unmappedCandidateOptions,
-  onSubmit,
-  onClose,
-}: ModalProps) {
-  const noMappingValue = 'noMapping';
-
-  const currentCandidate = !currentItem.candidateKey
-    ? undefined
-    : {
-        key: currentItem.candidateKey,
-        label: allCandidateOptions[currentItem.candidateKey].label,
-      };
-
-  const handleSubmit = ({ selectedCandidate }: FormValues) => {
-    onSubmit([
-      {
-        sourceKey: currentItem.sourceKey,
-        candidateKey:
-          selectedCandidate !== noMappingValue ? selectedCandidate : undefined,
-      },
-    ]);
-    onClose();
-  };
-
-  const options: RadioOption<string>[] = [
-    {
-      label: 'No mapping available',
-      value: noMappingValue,
-      divider: `Select ${itemType}:`,
-    },
-    ...(currentCandidate
-      ? [{ label: currentCandidate.label, value: currentCandidate.key }]
-      : []),
-    ...Object.entries(unmappedCandidateOptions).map(
-      ([candidateName, candidate]) => {
-        return {
-          label: candidate.label,
-          value: candidateName,
-        };
-      },
-    ),
-  ];
-
-  return (
-    <Modal open onExit={onClose} title={`Map existing ${itemType}`}>
-      <h3>Current data set {itemType}</h3>
-      <SummaryList>
-        <SummaryListItem term="Name">{currentItem?.sourceKey}</SummaryListItem>
-        <SummaryListItem term="Label">
-          {currentItem?.source.label}
-        </SummaryListItem>
-      </SummaryList>
-      <FormProvider
-        initialValues={{ selectedCandidate: currentItem.candidateKey }}
-        validationSchema={Yup.object({
-          selectedCandidate: Yup.string().required(
-            `Select the next data set ${itemType}`,
-          ),
-        })}
-      >
-        {({ formState }) => {
-          return (
-            <Form id="mapping-form" onSubmit={handleSubmit}>
-              <FormFieldRadioSearchGroup<FormValues>
-                alwaysShowOptions={[noMappingValue]}
-                hint={`Choose a ${itemType} that will be mapped to the current data set ${itemType} (see above).`}
-                legend={`Next data set ${itemType}`}
-                name="selectedCandidate"
-                options={options}
-                order={[]}
-                searchLabel={`Search ${itemType}s`}
-                small
-              />
-              <ButtonGroup>
-                <Button disabled={formState.isSubmitting} type="submit">
-                  {`Update ${itemType} mapping`}
-                </Button>
-                <ButtonText disabled={formState.isSubmitting} onClick={onClose}>
-                  Cancel
-                </ButtonText>
-              </ButtonGroup>
-            </Form>
-          );
-        }}
-      </FormProvider>
-    </Modal>
   );
 }
