+++ src/explore-education-statistics-admin/src/pages/release/data/components/__tests__/DataFilesReplacementTableRow.test.tsx
@@ -58,6 +58,7 @@ describe('DataFilesReplacementTableRow', () => {
       readyToPublish: false,
     },
     valid: true,
+    mapping: { indicators: { mappings: {}, candidates: {} } },
   };
 
   const defaultPermissions = {
