+++ src/explore-education-statistics-admin/src/pages/release/data/components/__tests__/ReleaseDataUploadsSection.test.tsx
@@ -127,6 +127,7 @@ describe('ReleaseDataUploadsSection', () => {
       valid: false,
       readyToPublish: false,
     },
+    mapping: { indicators: { mappings: {}, candidates: {} } },
   };
 
   beforeEach(() => {
